Add an documentation to the SimpleNestingMultiResource
This commit adds a documentation to SimpleNestingMultiResource as well
as modifies its deploy script to simply deploy it. User journey script
has been added to replace the showcase of SimpleNestingMultiResource
usage.

Custom script for running the SimpleNestingMultiResource user journey
has been added.

NestingMultiResource example smart contracts have been moved into their
own folder, so that the documentation of each architecture can be
separate from the main README.
