+++ package.json
@@ -7,10 +7,11 @@
     "deploy-equippable": "hardhat run scripts/deployEquippable.ts",
     "deploy-multi-resource": "hardhat run scripts/deployMultiResource.ts",
     "deploy-nesting": "hardhat run scripts/deployNesting.ts",
-    "deploy-nesting-multiresource": "hardhat run scripts/deployNestingMultiresource.ts",
+    "deploy-nesting-multi-resource": "hardhat run scripts/deployNestingMultiResource.ts",
     "deploy-split-equippable": "hardhat run scripts/deploySplitEquippable.ts",
     "user-journey-nesting": "hardhat run scripts/nestingUserJourney.ts",
-    "user-journey-multi-resource": "hardhat run scripts/multiResourceUserJourney.ts"
+    "user-journey-multi-resource": "hardhat run scripts/multiResourceUserJourney.ts",
+    "user-journey-nesting-multi-resource": "hardhat run scripts/nestingMultiResourceUserJourney.ts"
   },
   "keywords": [],
   "author": "@rmrk-team",
